requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.45
pydantic==2.12.5
PyJWT==2.9.0
cryptography==42.0.8
bcrypt==5.0.0
httpx==0.28.1
pypdf==6.6.0
python-dotenv==1.2.1
python-multipart==0.0.21
email-validator==2.3.0
